* VRPNDevice.java:  made the native_device member a long
	* *.cpp:  changed the handling of the native_device member
	from a java int to a java long.  this includes the get and
	set methods as well as the cached field descriptor.  cursory
	testing indicates that this works for at least one 32-bit
	environment.

+++ java_vrpn/vrpn/VRPNDevice.java
@@ -414,7 +414,7 @@ public abstract class VRPNDevice extends VRPN implements Runnable
 	// native device object (e.g., the vrpn_TrackerRemote, 
 	// vrpn_AnalogRemote, etc.).  this should be negative if the device 
 	// is uninitialized or has already been shut down
-	protected int native_device = -1;
+	protected long native_device = -1;
 	
 	// this is used to stop and to keep running the device thread
 	// in an orderly fashion.
