+++ java_vrpn/vrpn_TextReceiver.cpp
@@ -127,7 +127,7 @@ void VRPN_CALLBACK handle_text_message( void* userdata, const vrpn_TEXTCB info )
 JNIEXPORT void JNICALL 
 Java_vrpn_TextReceiver_mainloop( JNIEnv* env, jobject jobj )
 {
-  vrpn_Text_Receiver* t = (vrpn_Text_Receiver*) env->GetIntField( jobj, jfid_vrpn_VRPNDevice_native_device );
+  vrpn_Text_Receiver* t = (vrpn_Text_Receiver*) env->GetLongField( jobj, jfid_vrpn_VRPNDevice_native_device );
   if( t <= 0 )  // this text receiver is uninitialized or has been shut down already
     return;
   
@@ -175,8 +175,8 @@ Java_vrpn_TextReceiver_init( JNIEnv* env, jobject jobj, jstring jname,
   env->ReleaseStringUTFChars( jremoteOutLogfileName, remote_out_logfile_name );
   
   // now stash 't' in the jobj's 'native_device' field
-  jint jt = (jint) t;
-  env->SetIntField( jobj, jfid_vrpn_VRPNDevice_native_device, jt );
+  jlong jt = (jlong) t;
+  env->SetLongField( jobj, jfid_vrpn_VRPNDevice_native_device, jt );
   
 
 	return true;
@@ -189,7 +189,7 @@ JNIEXPORT void JNICALL
 Java_vrpn_TextReceiver_shutdownTextReceiver( JNIEnv* env, jobject jobj )
 {
   // get the tracker pointer
-  vrpn_Text_Receiver* t = (vrpn_Text_Receiver*) env->GetIntField( jobj, jfid_vrpn_VRPNDevice_native_device );
+  vrpn_Text_Receiver* t = (vrpn_Text_Receiver*) env->GetLongField( jobj, jfid_vrpn_VRPNDevice_native_device );
   
   // unregister a handler and destroy the text receiver
   if( t > 0 )
@@ -200,7 +200,7 @@ Java_vrpn_TextReceiver_shutdownTextReceiver( JNIEnv* env, jobject jobj )
   }
 
   // set the tracker pointer to -1
-  env->SetIntField( jobj, jfid_vrpn_VRPNDevice_native_device, -1 );
+  env->SetLongField( jobj, jfid_vrpn_VRPNDevice_native_device, -1 );
 
   // delete global reference to object (that was created in init)
   env->DeleteGlobalRef( jobj );
